rename: user repository 경로 수정
