--- src/main/java/com/burntoburn/easyshift/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.burntoburn.easyshift.repository;
-
-import com.burntoburn.easyshift.entity.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
